_scripts/requirements.txt
tweepy>=4.14
atproto>=0.0.55
Mastodon.py>=2.0
requests>=2.31
pyyaml>=6.0
